GetNextDispatch improvements for the client. improvements for the client.
Hiding the ugly memory details of the socket and native API by instantiating real classes before GetNextDispatch returns.  The caller only needs to switch on the EventId and cast to the proper Recv class type.

There is now a RegisterStructToDataDefinition call that goes with the AddToDataDefinition calls.  This enables the client library code to instantiate and populate the user struct before return.  Again hiding the ugly memory details.

Finally changed the SocketClient api to throw exceptions on HRESULT values that are errors.